backend/requirements.txt
flask>=3.0.0
flask-cors>=4.0.0
werkzeug>=3.0.0
requests>=2.31.0
paddlepaddle>=3.2.2
paddleocr>=3.3.0
python-dotenv>=1.0.0